requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
torch
torchvision
pillow
requests
transformers
safetensors
